First prototype of big notification
